+++ cmd/fpga_plugin/fpga_plugin.go
@@ -17,10 +17,14 @@ package main
 import (
 	"flag"
 	"fmt"
+	"io/ioutil"
 	"os"
+	"path"
+	"regexp"
+	"strings"
+	"time"
 
 	"github.com/golang/glog"
-	"golang.org/x/net/context"
 
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes"
@@ -29,8 +33,7 @@ import (
 	pluginapi "k8s.io/kubernetes/pkg/kubelet/apis/deviceplugin/v1beta1"
 	utilnode "k8s.io/kubernetes/pkg/util/node"
 
-	"github.com/intel/intel-device-plugins-for-kubernetes/cmd/fpga_plugin/devicecache"
-	"github.com/intel/intel-device-plugins-for-kubernetes/internal/deviceplugin"
+	dpapi "github.com/intel/intel-device-plugins-for-kubernetes/internal/deviceplugin"
 )
 
 const (
@@ -38,111 +41,276 @@ const (
 	devfsDirectory = "/dev"
 
 	// Device plugin settings.
-	pluginEndpointPrefix = "intel-fpga"
-	resourceNamePrefix   = "intel.com/fpga"
-	annotationName       = "com.intel.fpga.mode"
+	namespace      = "fpga.intel.com"
+	annotationName = "com.intel.fpga.mode"
+
+	// Scanner's mode of operation
+	afMode          = "af"
+	regionMode      = "region"
+	regionDevelMode = "regiondevel"
+
+	deviceRE = `^intel-fpga-dev.[0-9]+$`
+	portRE   = `^intel-fpga-port.[0-9]+$`
+	fmeRE    = `^intel-fpga-fme.[0-9]+$`
 )
 
-// deviceManager manages Intel FPGA devices.
-type deviceManager struct {
-	srv     deviceplugin.Server
-	fpgaID  string
-	name    string
-	mode    string
-	ch      chan map[string]deviceplugin.DeviceInfo
-	devices map[string]deviceplugin.DeviceInfo
-}
+type getDevTreeFunc func(devices []device) dpapi.DeviceTree
 
-func newDeviceManager(resourceName string, fpgaID string, mode string, ch chan map[string]deviceplugin.DeviceInfo) *deviceManager {
-	return &deviceManager{
-		fpgaID:  fpgaID,
-		name:    resourceName,
-		mode:    mode,
-		ch:      ch,
-		devices: make(map[string]deviceplugin.DeviceInfo),
-	}
-}
+// getRegionDevelTree returns mapping of region interface IDs to AF ports and FME devices
+func getRegionDevelTree(devices []device) dpapi.DeviceTree {
+	regionTree := dpapi.NewDeviceTree()
 
-func (dm *deviceManager) GetDevicePluginOptions(ctx context.Context, empty *pluginapi.Empty) (*pluginapi.DevicePluginOptions, error) {
-	fmt.Println("GetDevicePluginOptions: return empty options")
-	return new(pluginapi.DevicePluginOptions), nil
-}
-
-func (dm *deviceManager) sendDevices(stream pluginapi.DevicePlugin_ListAndWatchServer) error {
-	resp := new(pluginapi.ListAndWatchResponse)
-	for id, device := range dm.devices {
-		resp.Devices = append(resp.Devices, &pluginapi.Device{id, device.State})
-	}
-	glog.V(2).Info("Sending to kubelet ", resp.Devices)
-	if err := stream.Send(resp); err != nil {
-		dm.srv.Stop()
-		return fmt.Errorf("device-plugin: cannot update device list: %v", err)
-	}
-
-	return nil
-}
-
-// ListAndWatch returns a list of devices
-// Whenever a device state change or a device disappears, ListAndWatch returns the new list
-func (dm *deviceManager) ListAndWatch(empty *pluginapi.Empty, stream pluginapi.DevicePlugin_ListAndWatchServer) error {
-	glog.V(2).Info("Started ListAndWatch for ", dm.fpgaID)
-
-	if err := dm.sendDevices(stream); err != nil {
-		return err
-	}
-
-	for dm.devices = range dm.ch {
-		if err := dm.sendDevices(stream); err != nil {
-			return err
+	for _, dev := range devices {
+		for _, region := range dev.regions {
+			devType := fmt.Sprintf("%s-%s", regionMode, region.interfaceID)
+			devNodes := make([]string, len(region.afus)+1)
+			for num, afu := range region.afus {
+				devNodes[num] = afu.devNode
+			}
+			devNodes[len(region.afus)] = region.devNode
+			regionTree.AddDevice(devType, region.id, dpapi.DeviceInfo{
+				State: pluginapi.Healthy,
+				Nodes: devNodes,
+			})
 		}
 	}
 
-	return nil
+	return regionTree
 }
 
-func (dm *deviceManager) Allocate(ctx context.Context, rqt *pluginapi.AllocateRequest) (*pluginapi.AllocateResponse, error) {
-	response, err := deviceplugin.MakeAllocateResponse(rqt, dm.devices)
-	// Set container annotations when programming is allowed
-	if dm.mode == devicecache.RegionMode {
-		for _, containerResponse := range response.GetContainerResponses() {
-			containerResponse.Annotations = map[string]string{
-				annotationName: fmt.Sprintf("%s-%s", resourceNamePrefix, dm.mode),
+// getRegionTree returns mapping of region interface IDs to AF ports only
+func getRegionTree(devices []device) dpapi.DeviceTree {
+	regionTree := dpapi.NewDeviceTree()
+
+	for _, dev := range devices {
+		for _, region := range dev.regions {
+			devType := fmt.Sprintf("%s-%s", regionMode, region.interfaceID)
+			devNodes := make([]string, len(region.afus))
+			for num, afu := range region.afus {
+				devNodes[num] = afu.devNode
+			}
+			regionTree.AddDevice(devType, region.id, dpapi.DeviceInfo{
+				State: pluginapi.Healthy,
+				Nodes: devNodes,
+			})
+		}
+	}
+
+	return regionTree
+}
+
+// getAfuTree returns mapping of AFU IDs to AF ports
+func getAfuTree(devices []device) dpapi.DeviceTree {
+	afuTree := dpapi.NewDeviceTree()
+
+	for _, dev := range devices {
+		for _, region := range dev.regions {
+			for _, afu := range region.afus {
+				devType := fmt.Sprintf("%s-%s", afMode, afu.afuID)
+				afuTree.AddDevice(devType, afu.id, dpapi.DeviceInfo{
+					State: pluginapi.Healthy,
+					Nodes: []string{afu.devNode},
+				})
 			}
 		}
 	}
-	return response, err
+
+	return afuTree
 }
 
-func (dm *deviceManager) PreStartContainer(ctx context.Context, rqt *pluginapi.PreStartContainerRequest) (*pluginapi.PreStartContainerResponse, error) {
-	glog.Warning("PreStartContainer() should not be called")
-	return new(pluginapi.PreStartContainerResponse), nil
+type afu struct {
+	id      string
+	afuID   string
+	devNode string
 }
 
-func startDeviceManager(dm *deviceManager, pluginPrefix string) {
-	err := dm.srv.Serve(dm, dm.name, pluginPrefix)
+type region struct {
+	id          string
+	interfaceID string
+	devNode     string
+	afus        []afu
+}
+
+type device struct {
+	name    string
+	regions []region
+}
+
+type devicePlugin struct {
+	sysfsDir string
+	devfsDir string
+
+	deviceReg *regexp.Regexp
+	portReg   *regexp.Regexp
+	fmeReg    *regexp.Regexp
+
+	getDevTree      getDevTreeFunc
+	ignoreAfuIDs    bool
+	annotationValue string
+}
+
+// newDevicePlugin returns new instance of devicePlugin
+func newDevicePlugin(sysfsDir string, devfsDir string, mode string) (*devicePlugin, error) {
+	var getDevTree getDevTreeFunc
+
+	ignoreAfuIDs := false
+	annotationValue := ""
+	switch mode {
+	case afMode:
+		getDevTree = getAfuTree
+	case regionMode:
+		getDevTree = getRegionTree
+		ignoreAfuIDs = true
+		annotationValue = fmt.Sprintf("%s/%s", namespace, regionMode)
+	case regionDevelMode:
+		getDevTree = getRegionDevelTree
+		ignoreAfuIDs = true
+	default:
+		return nil, fmt.Errorf("Wrong mode: '%s'", mode)
+	}
+
+	return &devicePlugin{
+		sysfsDir:        sysfsDir,
+		devfsDir:        devfsDir,
+		deviceReg:       regexp.MustCompile(deviceRE),
+		portReg:         regexp.MustCompile(portRE),
+		fmeReg:          regexp.MustCompile(fmeRE),
+		getDevTree:      getDevTree,
+		ignoreAfuIDs:    ignoreAfuIDs,
+		annotationValue: annotationValue,
+	}, nil
+}
+
+func (dp *devicePlugin) PostAllocate(response *pluginapi.AllocateResponse) error {
+	// Set container annotations when programming is allowed
+	if len(dp.annotationValue) > 0 {
+		for _, containerResponse := range response.GetContainerResponses() {
+			containerResponse.Annotations = map[string]string{
+				annotationName: dp.annotationValue,
+			}
+		}
+	}
+
+	return nil
+}
+
+// Scan starts scanning of FPGA devices on the host
+func (dp *devicePlugin) Scan(notifier dpapi.Notifier) error {
+	for {
+		devTree, err := dp.scanFPGAs()
+		if err != nil {
+			glog.Error("Device scan failed: ", err)
+			return fmt.Errorf("Device scan failed: %v", err)
+		}
+
+		notifier.Notify(devTree)
+
+		time.Sleep(5 * time.Second)
+	}
+}
+
+func (dp *devicePlugin) getDevNode(devName string) (string, error) {
+	devNode := path.Join(dp.devfsDir, devName)
+	if _, err := os.Stat(devNode); err != nil {
+		return "", fmt.Errorf("Device %s doesn't exist: %v", devNode, err)
+	}
+
+	return devNode, nil
+}
+
+func (dp *devicePlugin) scanFPGAs() (dpapi.DeviceTree, error) {
+	var devices []device
+
+	glog.V(2).Info("Start new FPGA scan")
+
+	fpgaFiles, err := ioutil.ReadDir(dp.sysfsDir)
 	if err != nil {
-		glog.Fatal(err)
+		return nil, fmt.Errorf("Can't read sysfs folder: %v", err)
 	}
+
+	for _, fpgaFile := range fpgaFiles {
+		fname := fpgaFile.Name()
+
+		if !dp.deviceReg.MatchString(fname) {
+			continue
+		}
+
+		deviceFolder := path.Join(dp.sysfsDir, fname)
+		deviceFiles, err := ioutil.ReadDir(deviceFolder)
+		if err != nil {
+			return nil, err
+		}
+
+		regions, afus, err := dp.getSysFsInfo(deviceFolder, deviceFiles, fname)
+		if err != nil {
+			return nil, err
+		}
+
+		if len(regions) == 0 {
+			return nil, fmt.Errorf("No regions found for device %s", fname)
+		}
+
+		// Currently only one region per device is supported.
+		regions[0].afus = afus
+		devices = append(devices, device{
+			name:    fname,
+			regions: regions,
+		})
+	}
+
+	return dp.getDevTree(devices), nil
 }
 
-func handleUpdate(dms map[string]*deviceManager, updateInfo devicecache.UpdateInfo, start func(*deviceManager, string), mode string) {
-	glog.V(2).Info("Received dev updates: ", updateInfo)
-	for fpgaID, devices := range updateInfo.Added {
-		devCh := make(chan map[string]deviceplugin.DeviceInfo, 1)
-		resourceName := resourceNamePrefix + "-" + fpgaID
-		pPrefix := pluginEndpointPrefix + "-" + fpgaID
-		dms[fpgaID] = newDeviceManager(resourceName, fpgaID, mode, devCh)
-		go start(dms[fpgaID], pPrefix)
-		devCh <- devices
+func (dp *devicePlugin) getSysFsInfo(deviceFolder string, deviceFiles []os.FileInfo, fname string) ([]region, []afu, error) {
+	var regions []region
+	var afus []afu
+	for _, deviceFile := range deviceFiles {
+		name := deviceFile.Name()
+
+		if dp.fmeReg.MatchString(name) {
+			if len(regions) > 0 {
+				return nil, nil, fmt.Errorf("Detected more than one FPGA region for device %s. Only one region per FPGA device is supported", fname)
+			}
+			interfaceIDFile := path.Join(deviceFolder, name, "pr", "interface_id")
+			data, err := ioutil.ReadFile(interfaceIDFile)
+			if err != nil {
+				return nil, nil, err
+			}
+			devNode, err := dp.getDevNode(name)
+			if err != nil {
+				return nil, nil, err
+			}
+			regions = append(regions, region{
+				id:          name,
+				interfaceID: strings.TrimSpace(string(data)),
+				devNode:     devNode,
+			})
+		} else if dp.portReg.MatchString(name) {
+			var afuID string
+
+			if dp.ignoreAfuIDs {
+				afuID = "unused_afu_id"
+			} else {
+				afuFile := path.Join(deviceFolder, name, "afu_id")
+				data, err := ioutil.ReadFile(afuFile)
+				if err != nil {
+					return nil, nil, err
+				}
+				afuID = strings.TrimSpace(string(data))
+			}
+			devNode, err := dp.getDevNode(name)
+			if err != nil {
+				return nil, nil, err
+			}
+			afus = append(afus, afu{
+				id:      name,
+				afuID:   afuID,
+				devNode: devNode,
+			})
+		}
 	}
-	for fpgaID, devices := range updateInfo.Updated {
-		dms[fpgaID].ch <- devices
-	}
-	for fpgaID := range updateInfo.Removed {
-		dms[fpgaID].srv.Stop()
-		close(dms[fpgaID].ch)
-		delete(dms, fpgaID)
-	}
+
+	return regions, afus, nil
 }
 
 func main() {
@@ -154,8 +322,8 @@ func main() {
 
 	flag.StringVar(&kubeconfig, "kubeconfig", "", "absolute path to the kubeconfig file")
 	flag.StringVar(&master, "master", "", "master url")
-	flag.StringVar(&mode, "mode", string(devicecache.AfMode),
-		fmt.Sprintf("device plugin mode: '%s' (default), '%s' or '%s'", devicecache.AfMode, devicecache.RegionMode, devicecache.RegionDevelMode))
+	flag.StringVar(&mode, "mode", string(afMode),
+		fmt.Sprintf("device plugin mode: '%s' (default), '%s' or '%s'", afMode, regionMode, regionDevelMode))
 	flag.Parse()
 
 	if kubeconfig == "" {
@@ -185,9 +353,7 @@ func main() {
 		mode = nodeMode
 	}
 
-	updatesCh := make(chan devicecache.UpdateInfo)
-
-	cache, err := devicecache.NewCache(sysfsDirectory, devfsDirectory, mode, updatesCh)
+	plugin, err := newDevicePlugin(sysfsDirectory, devfsDirectory, mode)
 	if err != nil {
 		glog.Error(err)
 		os.Exit(1)
@@ -195,15 +361,6 @@ func main() {
 
 	glog.Info("FPGA device plugin started in ", mode, " mode")
 
-	go func() {
-		err := cache.Run()
-		if err != nil {
-			glog.Fatal(err)
-		}
-	}()
-
-	dms := make(map[string]*deviceManager)
-	for updateInfo := range updatesCh {
-		handleUpdate(dms, updateInfo, startDeviceManager, mode)
-	}
+	manager := dpapi.NewManager(namespace, plugin)
+	manager.Run()
 }
